requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
sentence-transformers>=2.2.2
ollama>=0.1.7
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0